# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: National Energy War Room — REST API
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Staff/citizen login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Citizen self-registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
...
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Logout
      responses:
        "200":
          description: Logged out
  /api/v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Current user
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /api/v1/complaints:
    post:
      operationId: fileComplaint
      tags: [complaints]
      summary: File a new complaint (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileComplaintRequest"
      responses:
        "201":
          description: Complaint filed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintFiledResponse"
  /api/v1/complaints/track/{ticketNumber}:
    get:
      operationId: trackComplaint
      tags: [complaints]
      summary: Track complaint by ticket number (requires phone last-4)
      parameters:
        - name: ticketNumber
          in: path
          required: true
          schema:
            type: string
        - name: phoneLast4
          in: query
          required: true
          schema:
            type: string
            minLength: 4
            maxLength: 4
      responses:
        "200":
          description: Complaint details
...
  /api/v1/admin/complaints:
    get:
      operationId: adminListComplaints
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: discoId
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: escalationLevel
          in: query
          schema:
            type: integer
        - name: slaBreached
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Complaints list
...
  /api/v1/admin/complaints/{id}:
    get:
      operationId: adminGetComplaint
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complaint detail
    patch:
      operationId: updateComplaint
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated complaint
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, fullName, phone]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        phone:
          type: string
        nin:
          type: string
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            user:
              $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        fullName:
          type: string
        organisationId:
          type: string
    FileComplaintRequest:
      type: object
      required: [citizenName, citizenPhone, discoId, category, description]
      properties:
        citizenName:
          type: string
        citizenPhone:
          type: string
        citizenEmail:
          type: string
        discoId:
          type: string
        feederId:
          type: string
        category:
          type: string
        description:
          type: string
        location:
          type: string
    ComplaintFiledResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticketNumber:
              type: string
            id:
              type: string
            status:
              type: string
            message:
              type: string
...